Correct prefix notation
I think I have managed to get the correct prefix notation for a sympy formula.
The main problem was that multiplication and addition in sympy do not have a fixed number of arguments,
but variable one, so mul a b c would be a*b*c, but I want mul a mul b c.
Took some inspiration from https://github.com/facebookresearch/SymbolicMathematics,
like the list for the operators, since I could not find it in sympy.