Fixed keras tuner save problem
Keras tuner was complaining that it could not save the model.
First I thought that the way the transformer was built, was the problem,
but it turned out, that I had mixed `tensorflow.keras` and `keras`.
Fully going to `tensorflow.keras` fixed it.

Also changed on colab, where it is currently running.